services/pdf-generator-service/requirements.txt
flask==3.0.0
reportlab==4.0.7
weasyprint==60.2
jinja2==3.1.2
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
requests==2.31.0
gunicorn==21.2.0
pillow==10.1.0
matplotlib==3.8.2